# frontend/src/types/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthAssistantBundleImportRequest": {
        "properties": {
          "bind_matching_sources": {
            "default": true,
            "title": "Bind Matching Sources",
            "type": "boolean"
          },
          "bundle": {
            "additionalProperties": true,
            "title": "Bundle",
            "type": "object"
          },
          "create_browser_session": {
            "default": true,
            "title": "Create Browser Session",
            "type": "boolean"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "required": [
          "bundle"
        ],
        "title": "AuthAssistantBundleImportRequest",
        "type": "object"
      },
...
      "Body_import_auth_export_zip_from_assistant_api_auth_assistant_auth_exports_import_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_import_auth_export_zip_from_assistant_api_auth_assistant_auth_exports_import_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PairingClaimRequest": {
        "properties": {
          "app_version": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Version"
          },
          "device_name": {
            "default": "PIM Auth Assistant",
            "maxLength": 255,
            "minLength": 1,
            "title": "Device Name",
            "type": "string"
          },
          "pairing_token": {
            "title": "Pairing Token",
            "type": "string"
          }
        },
        "required": [
          "pairing_token"
        ],
        "title": "PairingClaimRequest",
        "type": "object"
      },
      "PairingClaimResponse": {
        "properties": {
          "capabilities": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Capabilities",
            "type": "object"
          },
          "device": {
            "additionalProperties": true,
            "title": "Device",
            "type": "object"
          },
          "device_token": {
            "title": "Device Token",
            "type": "string"
          },
          "server": {
            "additionalProperties": true,
            "title": "Server",
            "type": "object"
          }
        },
        "required": [
          "device_token",
          "device",
          "server",
          "capabilities"
        ],
        "title": "PairingClaimResponse",
        "type": "object"
      },
      "PairingTokenCreateRequest": {
        "properties": {
          "ttl_minutes": {
            "default": 10,
            "maximum": 60.0,
            "minimum": 1.0,
            "title": "Ttl Minutes",
            "type": "integer"
          }
        },
        "title": "PairingTokenCreateRequest",
        "type": "object"
      },
      "PairingTokenResponse": {
        "properties": {
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "pairing_token": {
            "title": "Pairing Token",
            "type": "string"
          },
          "pairing_url": {
            "title": "Pairing Url",
            "type": "string"
          },
          "token_hint": {
            "title": "Token Hint",
            "type": "string"
          }
        },
        "required": [
          "pairing_token",
          "pairing_url",
          "expires_at",
          "token_hint"
        ],
        "title": "PairingTokenResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "个人化资讯监控管理系统 API",
    "title": "Personal Information Monitor",
    "version": "1.4.3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth-assistant/auth-bundles/import": {
      "post": {
        "operationId": "import_auth_bundle_from_assistant_api_auth_assistant_auth_bundles_import_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthAssistantBundleImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Import Auth Bundle From Assistant",
        "tags": [
          "auth-assistant"
        ]
      }
    },
    "/api/auth-assistant/auth-exports/import": {
      "post": {
        "operationId": "import_auth_export_zip_from_assistant_api_auth_assistant_auth_exports_import_post",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_auth_export_zip_from_assistant_api_auth_assistant_auth_exports_import_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Import Auth Export Zip From Assistant",
        "tags": [
          "auth-assistant"
        ]
      }
    },
    "/api/auth-assistant/devices": {
      "get": {
        "operationId": "list_devices_api_auth_assistant_devices_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Devices",
        "tags": [
          "auth-assistant"
        ]
      }
    },
    "/api/auth-assistant/devices/{device_id}": {
      "delete": {
        "operationId": "revoke_device_api_auth_assistant_devices__device_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "title": "Device Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Revoke Device",
        "tags": [
          "auth-assistant"
        ]
      }
    },
    "/api/auth-assistant/pair": {
      "post": {
        "description": "Exchange a short one-time pairing token for a device token.",
        "operationId": "claim_pairing_token_api_auth_assistant_pair_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingClaimResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Pairing Token",
        "tags": [
          "auth-assistant"
        ]
      }
    },
    "/api/auth-assistant/pairing-tokens": {
      "post": {
        "description": "Create a one-time pairing token from authenticated PIM Web.",
        "operationId": "create_pairing_token_api_auth_assistant_pairing_tokens_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairingTokenCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairingTokenResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Pairing Token",
        "tags": [
          "auth-assistant"
        ]
      }
    },
...
  }
}